Don't include files that come from indirect NuGet dependencies
This fixes issue #10 which causes the following scenario to not work.
Consider this project structure:

  A.csproj => ANuGetPackage
  A.nuproj => A.csproj
  B.csproj => A.csproj
  B.nuproj => B.csproj, A.nuproj

B.nupkg should not contain assemblies from ANuGetPackage but it does.
The root cause is that we don't use the closure of NuGet packages to
exclude files where we should.

The current behavior will now also cause the package to depend on the
closure of packages, as opposed to only the direct dependencies. We
should probably fix that.